Refactor custom mint module functions (#31)
The current implementation uses `if conditions` to check for chain which have custom mint modules and then follows a different set of operations. 
Going forward there might be many such chains which will implement custom modules and the "if" approach might not be the optimal way to go.
I have considered a map[string]func() which maps a chain name to the required functions.

Please do suggest if you have any better ideas to solve this.